# ai-reply-back-end/docs/openapi.yaml
openapi: 3.0.3
info:
  title: AI Reply Backend API
  version: "1.0.0"
  description: >
    Мобильный API AI Reply. Тексты сообщений не сохраняются: сервер обрабатывает
    их только в памяти во время запроса.
externalDocs:
  description: Публичная оферта
  url: https://api.meily.kz/offer
servers:
  - url: https://api.meily.kz
    description: production
paths:
  /healthz:
    get:
      summary: Живость сервиса
      responses: { "200": { description: OK } }
  /api/v1/config:
    get:
      summary: Публичная конфигурация клиента (языки, лимиты, страны)
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: { $ref: "#/components/schemas/PublicConfig" }
  /api/v1/plans:
    get:
      summary: Активные тарифы
      responses: { "200": { description: OK } }
  /api/v1/auth/request-otp:
    post:
      summary: Запросить код подтверждения
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [identifier]
              properties:
                identifier: { type: string, example: "+7 701 123 45 67" }
                locale: { type: string, enum: [kk, ru, en, uz] }
      responses:
        "200": { description: Запрос принят, код отправлен доступным каналом }
        "400": { description: INVALID_REQUEST — номер не поддерживается }
        "429": { description: RATE_LIMITED }
  /api/v1/auth/verify-otp:
    post:
      summary: Подтвердить код и получить сессию
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [identifier, code]
              properties:
                identifier: { type: string }
                code: { type: string, example: "4827" }
                device:
                  type: object
                  properties:
                    device_id: { type: string }
                    platform: { type: string, enum: [ios, android, web] }
                    app_version: { type: string }
                    os_version: { type: string }
                    locale: { type: string }
                    timezone: { type: string }
      responses:
        "200": { description: access_token, refresh_token, user, usage }
        "400": { description: INVALID_OTP / OTP_EXPIRED }
  /api/v1/auth/refresh:
    post:
      summary: Обновить пару токенов (с ротацией)
      responses:
        "200": { description: OK }
        "401": { description: UNAUTHORIZED }
  /api/v1/auth/logout:
    post:
      summary: Отозвать refresh-токен
      responses: { "200": { description: OK } }
  /api/v1/me:
    get:
      summary: Профиль, тариф и квота
      security: [{ bearerAuth: [] }]
      responses: { "200": { description: OK } }
    patch:
      summary: Частичное обновление профиля
      security: [{ bearerAuth: [] }]
      responses: { "200": { description: OK } }
  /api/v1/me/usage:
    get:
      summary: Остаток генераций
      security: [{ bearerAuth: [] }]
      responses: { "200": { description: OK } }
  /api/v1/me/subscription:
    get:
      summary: Текущая подписка
      security: [{ bearerAuth: [] }]
      responses: { "200": { description: OK } }
  /api/v1/me/consents:
    post:
      summary: Сохранить принятие текущей оферты и политики конфиденциальности
      security: [{ bearerAuth: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/LegalConsentRequest" }
      responses:
        "200":
          description: Согласие сохранено
          content:
            application/json:
              schema: { $ref: "#/components/schemas/LegalConsent" }
        "409": { description: LEGAL_VERSION_MISMATCH — клиент должен обновить конфигурацию }
  /api/v1/devices:
    post:
      summary: Регистрация устройства и push-токена
      security: [{ bearerAuth: [] }]
      responses: { "200": { description: OK } }
  /api/v1/devices/{id}:
    delete:
      summary: Отозвать устройство
      security: [{ bearerAuth: [] }]
      parameters:
        - { name: id, in: path, required: true, schema: { type: string } }
      responses: { "200": { description: OK } }
  /api/v1/ai/reply:
    post:
      summary: Сгенерировать ответ
      security: [{ bearerAuth: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [source_text]
              properties:
                source_text: { type: string, maxLength: 300 }
                instruction: { type: string, maxLength: 400 }
                language: { type: string, enum: [kk, ru, en, uz] }
                template_id: { type: string }
      responses:
        "200": { description: reply + usage }
        "429": { description: DAILY_LIMIT_REACHED }
        "502": { description: AI_PROVIDER_UNAVAILABLE }
        "504": { description: AI_TIMEOUT }
  /v1/auth/register:
    post:
      summary: "Совместимость: install_id → токен (старые сборки)"
      deprecated: true
      responses: { "200": { description: OK } }
  /v1/reply/generate:
    post:
      summary: "Совместимость: прежний формат генерации"
      deprecated: true
      responses: { "200": { description: OK } }
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    LegalConfig:
      type: object
      required: [terms_version, privacy_version, terms_url, privacy_url]
      properties:
        terms_version: { type: string, example: "2026-09-19" }
        privacy_version: { type: string, example: "2026-09-19" }
        terms_url: { type: string, format: uri, example: "https://api.meily.kz/offer" }
        privacy_url: { type: string, format: uri, example: "https://api.meily.kz/privacy" }
    PublicConfig:
      type: object
      properties:
        locales:
          type: array
          items: { type: string, enum: [kk, ru, en, uz] }
        timezone: { type: string, example: "Asia/Almaty" }
        max_source_characters: { type: integer, example: 300 }
        max_instruction_length: { type: integer, example: 400 }
        legal: { $ref: "#/components/schemas/LegalConfig" }
    LegalConsentRequest:
      type: object
      required: [terms_version, privacy_version, locale, platform]
      properties:
        terms_version: { type: string }
        privacy_version: { type: string }
        locale: { type: string, enum: [kk, ru, en, uz] }
        platform: { type: string, enum: [ios, android] }
        app_version: { type: string }
    LegalConsent:
      allOf:
        - { $ref: "#/components/schemas/LegalConsentRequest" }
        - type: object
          required: [accepted_at]
          properties:
            accepted_at: { type: string, format: date-time }